third_party/blink/web_tests/virtual/saa-beyond-cookies-disabled/external/wpt/storage-access-api/storage-access-beyond-cookies.none.tentative.sub.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] Verify StorageAccessAPIBeyondCookies for None
  assert_equals: Storage Access API should not allow access for empty requests. expected "HasAccess for none" but got "Unable to load handle in cross-site context for all"
Harness: the test ran to completion.
